Ignore SIGPIPE
If stdout ever gets closed, a SIGPIPE is received which immediately
terminates the program.  This could happen if GCF's output is being
piped to another program (e.g., tee) and then the user sends a SIGINT to
terminate the process group.  In this case, the pipe would be closed and
GCF would receive a SIGPIPE.

By ignoring SIGPIPE, we allow the SIGINT handler to finish running and
clean up any current workers.  All writes to stdout (i.e. logging) will
fail, so we won't be able to tell from the output whether cleanup was
successful.  But from testing manually, all workers were properly shut
down.